Fix CI test failures by updating function imports and test data
- Update test imports to match actual function names in walker.py
- Fix test_save_puzzle_to_csv.py: save_puzzle_to_csv → save_position_to_csv
- Fix test_walker.py: create_puzzle_data → create_position_data, generate_and_save_puzzles → generate_and_save_positions
- Fix test_divergence.py: adjust test data to meet MIN_WIN_RATE_DELTA threshold
- Fix pandas FutureWarning by explicitly setting dtype=float in contingency DataFrame

+++ src/divergence.py
@@ -44,7 +44,7 @@ def check_frequency_divergence(
     Returns: tuple[bool, float]: A tuple containing a boolean indicating if there is a significant difference in move frequencies and the p-value of the chi-square test.
     """
     all_moves = set(base_df["Move"]).union(set(target_df["Move"]))
-    contingency = pd.DataFrame(index=list(all_moves), columns=["Base", "Target"]).fillna(0)
+    contingency = pd.DataFrame(index=list(all_moves), columns=["Base", "Target"], dtype=float).fillna(0)
     for move in all_moves:
         base_games = base_df[base_df["Move"] == move]["Games"].sum() if move in base_df["Move"].values else 0
         target_games = target_df[target_df["Move"] == move]["Games"].sum() if move in target_df["Move"].values else 0
